+++ pyfakefs/pytest_tests/pytest_fixture_test.py
@@ -13,6 +13,7 @@
 # Example for a test using a custom pytest fixture with an argument to Patcher
 
 import pytest
+import undefined
 
 import pyfakefs.pytest_tests.example as example
 from pyfakefs.fake_filesystem_unittest import Patcher
@@ -39,6 +40,12 @@ def test_example_file_passing_using_patcher():
     with Patcher(modules_to_reload=[example]) as patcher:
         patcher.fs.create_file(example.EXAMPLE_FILE, contents="stuff here")
         check_that_example_file_is_in_fake_fs()
+
+
+def test_undefined(fs):
+    # regression test for #923
+    with pytest.raises(NotImplementedError):
+        print(undefined)
 
 
 def check_that_example_file_is_in_fake_fs():